% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R, R/build_shiny.R
\name{build_shiny}
\alias{build_shiny}
\title{Build Shiny app}
\usage{
build_shiny(spec = NULL, ...)

build_shiny(spec = NULL, ...)
}
\arguments{
\item{spec}{Path to a YAML spec (e.g., 'inst/templates/zashboard.yml') or a list-like spec already parsed. Defaults to NULL.}

\item{...}{Additional options passed to the Shiny builder.}
}
\description{
Build Shiny app

Build a Shiny app
}
