% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R, R/build_static.R
\name{build_static}
\alias{build_static}
\title{Build static HTML}
\usage{
build_static(spec = NULL, ...)

build_static(spec = NULL, ...)
}
\arguments{
\item{spec}{Path to a YAML spec (e.g., 'inst/templates/zashboard.yml') or a list-like spec already parsed. Defaults to NULL.}

\item{...}{Additional options passed to the static builder.}
}
\description{
Build static HTML

Build static HTML output
}
