% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectors.R
\name{zashboard_disconnect}
\alias{zashboard_disconnect}
\title{Disconnect a zashboard connector}
\usage{
zashboard_disconnect(conn)
}
\arguments{
\item{conn}{A zashboard connection object.}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Calls the embedded \code{disconnect} function if present.
}
