% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectors.R
\name{zashboard_execute_collect}
\alias{zashboard_execute_collect}
\title{Collect rows from an Arrow dataset connector}
\usage{
zashboard_execute_collect(conn)
}
\arguments{
\item{conn}{A zashboard Arrow connection from \code{zashboard_connect_arrow()}.}
}
\value{
A data frame with all rows.
}
\description{
Collect rows from an Arrow dataset connector
}
