% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_connectors.R
\name{zashboard_execute_sql}
\alias{zashboard_execute_sql}
\title{Execute a SQL query on a connector (DuckDB/MSSQL)}
\usage{
zashboard_execute_sql(conn, sql)
}
\arguments{
\item{conn}{A zashboard connection object from \code{zashboard_connect_*()}.}

\item{sql}{A single SQL string.}
}
\value{
A data frame with results.
}
\description{
Execute a SQL query on a connector (DuckDB/MSSQL)
}
