% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{zashboard_format_errors}
\alias{zashboard_format_errors}
\title{Format a vector of error messages}
\usage{
zashboard_format_errors(errs)
}
\arguments{
\item{errs}{character vector}
}
\description{
Format a vector of error messages
}
\keyword{internal}
