% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R, R/build_quarto.R
\name{build_quarto}
\alias{build_quarto}
\title{Build Quarto}
\usage{
build_quarto(spec = NULL, ...)

build_quarto(spec = NULL, ...)
}
\arguments{
\item{spec}{Path to a YAML spec (e.g., 'inst/templates/zashboard.yml') or a list-like spec already parsed. Defaults to NULL.}

\item{...}{Additional options passed to the Quarto builder.}
}
\description{
Build Quarto

Build a Quarto site for the dashboard
}
