% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R, R/build_shinylive.R
\name{build_shinylive}
\alias{build_shinylive}
\title{Build Shinylive app}
\usage{
build_shinylive(spec = NULL, ...)

build_shinylive(spec = NULL, ...)
}
\arguments{
\item{spec}{Path to a YAML spec (e.g., 'inst/templates/zashboard.yml') or a list-like spec already parsed. Defaults to NULL.}

\item{...}{Additional options passed to the Shinylive builder.}
}
\description{
Build Shinylive app

Build a Shinylive app
}
