% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R, R/build_static.R
\name{build_static}
\alias{build_static}
\title{Build static HTML}
\usage{
build_static(
  spec,
  out_dir = file.path(tempdir(), "zashboard-static"),
  overwrite = TRUE,
  title = NULL,
  ...
)

build_static(
  spec,
  out_dir = file.path(tempdir(), "zashboard-static"),
  overwrite = TRUE,
  title = NULL,
  ...
)
}
\arguments{
\item{spec}{Path to YAML spec (or spec list)}

\item{out_dir}{Output directory}

\item{overwrite}{Overwrite output directory if it exists}

\item{title}{Optional page title override}
}
\description{
Build static HTML

Build a static HTML dashboard from a spec
}
