% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{zashboard-package}
\alias{zashboard}
\alias{zashboard-package}
\title{Zashboard}
\description{
Build dashboards from a single spec to static HTML, Shiny, or Shinylive.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zobaer09/zashboard}
  \item \url{https://zobaer09.github.io/zashboard/}
  \item Report bugs at \url{https://github.com/zobaer09/zashboard/issues}
}

}
\author{
\strong{Maintainer}: Zobaer Ahmed \email{zobaer09@users.noreply.github.com}

}
\keyword{internal}
