% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{zashboard_as_manifest}
\alias{zashboard_as_manifest}
\title{Convert a Zashboard spec to a manifest}
\usage{
zashboard_as_manifest(spec = NULL)
}
\arguments{
\item{spec}{Path to a YAML spec file \strong{or} a pre-parsed list with the same
structure. If \code{NULL}, the package template spec is used.}
}
\value{
A manifest list suitable for downstream build targets.
}
\description{
Convert a Zashboard spec to a manifest
}
