% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{zashboard_manifest}
\alias{zashboard_manifest}
\title{Create a Zashboard manifest (convenience wrapper)}
\usage{
zashboard_manifest(spec = NULL)
}
\arguments{
\item{spec}{Path to a YAML spec file \strong{or} a pre-parsed list with the same
structure. If \code{NULL}, the package template spec is used.}
}
\value{
A manifest list.
}
\description{
Create a Zashboard manifest (convenience wrapper)
}
