% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R, R/spec_io.R
\name{zashboard_read_spec}
\alias{zashboard_read_spec}
\title{Read a Zashboard spec (YAML) from disk}
\usage{
zashboard_read_spec(spec = NULL)

zashboard_read_spec(spec = NULL)
}
\arguments{
\item{spec}{Path to a YAML spec file \strong{or} a pre-parsed list with the same
structure. If \code{NULL}, the package template spec is used.}

\item{path}{Path to a YAML file}
}
\value{
A list representing the spec

A validated spec (list).
}
\description{
Read a Zashboard spec (YAML) from disk

Read and validate a Zashboard spec
}
